Test case for decimal parsing

+++ Tests/ParserTests/DictionaryExtensionTests.swift
@@ -11,7 +11,7 @@ final class DictionaryExtensionTests: XCTestCase {
                 "a": 1,
                 "b": "2",
                 "c": {
-                    "d": 3
+                    "d": 3.0
                 },
                 "f": 2,
                 "h": {
@@ -46,7 +46,7 @@ final class DictionaryExtensionTests: XCTestCase {
                 "a": 1,
                 "b": 4,
                 "c": {
-                    "d": 3,
+                    "d": 3.0,
                     "e": "5"
                 },
                 "f": {
@@ -69,34 +69,34 @@ final class DictionaryExtensionTests: XCTestCase {
         let dict2 = try decoder.decode(AnyParsable.self, from: json2).value as? [String: AnyParsable] ?? [:]
         let dict3 = try decoder.decode(AnyParsable.self, from: json3).value as? [String: AnyParsable] ?? [:]
         try context("dict1") {
-            XCTAssertEqual(dict1["a"]?.value as? Int, 1)
+            XCTAssertEqual(dict1["a"]?.value as? Decimal, 1)
             XCTAssertEqual(dict1["b"]?.value as? String, "2")
 
             let c = try XCTUnwrap(dict1["c"]?.value as? [String: AnyParsable])
-            XCTAssertEqual(c["d"]?.value as? Int, 3)
+            XCTAssertEqual(c["d"]?.value as? Decimal, 3.0)
             XCTAssertNil(c["e"]?.value)
-            XCTAssertEqual(dict1["f"]?.value as? Int, 2)
+            XCTAssertEqual(dict1["f"]?.value as? Decimal, 2)
 
             let h = try XCTUnwrap(dict1["h"]?.value as? [String: AnyParsable])
             let i = try XCTUnwrap(h["i"]?.value as? [String: AnyParsable])
-            XCTAssertEqual(i["j"]?.value as? Int, 9)
+            XCTAssertEqual(i["j"]?.value as? Decimal, 9)
             XCTAssertEqual(i["k"]?.value as? Bool, true)
 
             let l = try XCTUnwrap(i["l"]?.value as? [AnyParsable])
-            XCTAssertEqual(l.map { $0.value as? Int }, [1, 2, 3])
+            XCTAssertEqual(l.map { $0.value as? Decimal }, [1, 2, 3])
             XCTAssertNil(i["m"]?.value)
-            XCTAssertEqual(h["n"]?.value as? Int, 0)
+            XCTAssertEqual(h["n"]?.value as? Decimal, 0)
         }
         try context("dict2") {
             XCTAssertNil(dict2["a"]?.value)
-            XCTAssertEqual(dict2["b"]?.value as? Int, 4)
+            XCTAssertEqual(dict2["b"]?.value as? Decimal, 4)
 
             let c = try XCTUnwrap(dict2["c"]?.value as? [String: AnyParsable])
             XCTAssertNil(c["d"]?.value)
             XCTAssertEqual(c["e"]?.value as? String, "5")
 
             let f = try XCTUnwrap(dict2["f"]?.value as? [String: AnyParsable])
-            XCTAssertEqual(f["g"]?.value as? Int, 6)
+            XCTAssertEqual(f["g"]?.value as? Decimal, 6)
 
             let h = try XCTUnwrap(dict2["h"]?.value as? [String: AnyParsable])
 
@@ -105,31 +105,31 @@ final class DictionaryExtensionTests: XCTestCase {
             XCTAssertNil(i["k"]?.value)
 
             let l = try XCTUnwrap(i["l"]?.value as? [AnyParsable])
-            XCTAssertEqual(l.map { $0.value as? Int }, [4])
-            XCTAssertEqual(i["m"]?.value as? Int, 7)
+            XCTAssertEqual(l.map { $0.value as? Decimal }, [4])
+            XCTAssertEqual(i["m"]?.value as? Decimal, 7)
             XCTAssertNil(h["n"]?.value)
         }
         try context("dict3") {
-            XCTAssertEqual(dict3["a"]?.value as? Int, 1)
-            XCTAssertEqual(dict3["b"]?.value as? Int, 4)
+            XCTAssertEqual(dict3["a"]?.value as? Decimal, 1)
+            XCTAssertEqual(dict3["b"]?.value as? Decimal, 4)
 
             let c = try XCTUnwrap(dict3["c"]?.value as? [String: AnyParsable])
-            XCTAssertEqual(c["d"]?.value as? Int, 3)
+            XCTAssertEqual(c["d"]?.value as? Decimal, 3.0)
             XCTAssertEqual(c["e"]?.value as? String, "5")
 
             let f = try XCTUnwrap(dict3["f"]?.value as? [String: AnyParsable])
-            XCTAssertEqual(f["g"]?.value as? Int, 6)
+            XCTAssertEqual(f["g"]?.value as? Decimal, 6)
 
             let h = try XCTUnwrap(dict3["h"]?.value as? [String: AnyParsable])
 
             let i = try XCTUnwrap(h["i"]?.value as? [String: AnyParsable])
-            XCTAssertEqual(i["j"]?.value as? Int, 9)
+            XCTAssertEqual(i["j"]?.value as? Decimal, 9)
             XCTAssertEqual(i["k"]?.value as? Bool, true)
 
             let l = try XCTUnwrap(i["l"]?.value as? [AnyParsable])
-            XCTAssertEqual(l.map { $0.value as? Int }, [1, 2, 3, 4])
-            XCTAssertEqual(i["m"]?.value as? Int, 7)
-            XCTAssertEqual(h["n"]?.value as? Int, 0)
+            XCTAssertEqual(l.map { $0.value as? Decimal }, [1, 2, 3, 4])
+            XCTAssertEqual(i["m"]?.value as? Decimal, 7)
+            XCTAssertEqual(h["n"]?.value as? Decimal, 0)
         }
         context("dict1 + dict2 equal dict3") {
             XCTAssertEqual(AnyParsable(dict1 + dict2), AnyParsable(dict3))
